Citas y afinar detalles

+++ src/routes/routes.js
@@ -28,8 +28,25 @@ const { fstat } = require('fs');
 
 //------------------------------------------------------------------------------------------------
 
-router.get('/', (req, res, next) => {
-	
+router.get('/', async(req, res, next) => {
+	const usuarios = await Usuarios.findOne({ tipo: "Admin" });
+	if(!usuarios){
+		const admin = await new Personas();
+		const user = await new Usuarios();
+		admin.nombre = "Admin-Nombre";
+		admin.apellido = "Admin-Apellido";
+		admin.nacimiento = "2001-01-01";
+		admin.cedula = "";
+		admin.telefono = "";
+		admin.urlimage = "https://res.cloudinary.com/djsa7v6bs/image/upload/v1643855173/perfil_pzarxl.jpg";
+		admin.public_id = "perfil_pzarxl";
+		user.email = "admin@admin.com";
+		user.password = "admin";
+		user.tipo = "Admin";
+		admin.usuario = user._id;
+		user.save();
+		admin.save();
+	}
 	res.render('home');
 });
 
@@ -88,6 +105,9 @@ router.post('/perfil', isAuthenticated, async (req, res, next) => {
 	persona.apellido = apellido;
 	persona.nacimiento = nacimiento;
 	persona.telefono = telefono;
+	if(usuario.tipo == "Admin"){
+		persona.cedula = req.body.cedula;
+	}
 	usuario.email = email;
 	usuario.password = password;
 	if (req.file) {
@@ -99,7 +119,9 @@ router.post('/perfil', isAuthenticated, async (req, res, next) => {
 	}
 	await persona.save();
 	await usuario.save();
-	await unlink(req.file.path);
+	if(req.file){
+		await unlink(req.file.path);
+	}
 	res.render('perfil', {
 		persona: persona,
 		edad: getEdad(persona.nacimiento)
@@ -107,32 +129,101 @@ router.post('/perfil', isAuthenticated, async (req, res, next) => {
 });
 
 router.get('/reservar', isAuthenticated, async (req, res, next) => {
-	const servicios = await Servicios.find({});
+	const servicios = await Servicios.find({ tipo: "Fijo" });
 	res.render('reservar', {
 		ServiciosLista: servicios
 	});
 });
 
 router.post('/reservar', isAuthenticated, async (req, res, next) => {
-
-	// const servicios = await Servicios.find({});
-	// res.render('reservar', {
-	// 	ServiciosLista: servicios
-	// });
+	const servicios = await Servicios.find({ tipo: "Fijo" });
+	const { btnRadio } = req.body;
+	const userEmpleados = await Usuarios.find({ tipo: "Empleado" });
+	const aux = await Usuarios.find({ tipo: "Empleado" });
+	const ids = [];
+	aux.forEach(auxi => {
+		ids.push(auxi._id);
+	});
+	const personaEmpleados = await Personas.find({ usuario: { $in: ids } });
+	localStorage.setItem("selServicio", JSON.stringify(servicios[parseInt(btnRadio)]));
+	res.render('seleccionar-empleado', {
+		userEmpleados: userEmpleados,
+		personaEmpleados: personaEmpleados,
+		selServicios: servicios[parseInt(btnRadio)]
+	});
 });
 
+router.post('/seleccionar-empleado', isAuthenticated, async (req, res, next) => {
 
+	const aux = await Usuarios.find({ tipo: "Empleado" });
+	const ids = [];
+	aux.forEach(auxi => {
+		ids.push(auxi._id);
+	});
+	const personaEmpleados = await Personas.find({ usuario: { $in: ids } });
 
-router.get('/seleccionar-empleado', isAuthenticated, (req, res, next) => {
-	res.render('seleccionar-empleado');
+	const { btnradio, fecha, hora } = req.body;
+	const servicio = JSON.parse(localStorage.getItem("selServicio"));
+	const newCita = new Citas();
+	newCita.fechaCita = fecha;
+	newCita.horaCita = hora;
+	newCita.estado = "Pendiente";
+	const cliente = await Personas.findOne({ usuario: req.user._id });
+	console.log(cliente);
+	newCita.cliente = cliente._id;
+	newCita.servicio = servicio._id;
+	newCita.empleado = personaEmpleados[parseInt(btnradio)]._id;
+	newCita.save();
+	res.render('home');
 });
+
 
 router.get('/personalizar', isAuthenticated, (req, res, next) => {
 	res.render('personalizar');
 });
 
-router.get('/admin/cliente', isAuthenticated, (req, res, next) => {
-	res.render('admin/cliente');
+router.post('/personalizar', isAuthenticated, async (req, res, next) => {
+	const servicios = new Servicios();
+	const { nombre, descripcion } = req.body;
+	servicios.nombre = nombre;
+	servicios.descripcion = descripcion;
+	servicios.tipo = "Personalizado";
+	servicios.precio = 0.00;
+	const result = await Cloudinary.v2.uploader.upload(req.file.path);
+	servicios.urlimage = result.url;
+	servicios.public_id = result.public_id;
+	await servicios.save();
+
+	const userEmpleados = await Usuarios.find({ tipo: "Empleado" });
+	const aux = await Usuarios.find({ tipo: "Empleado" });
+	const ids = [];
+	aux.forEach(auxi => {
+		ids.push(auxi._id);
+	});
+
+	const personaEmpleados = await Personas.find({ usuario: { $in: ids } });
+	localStorage.setItem("selServicio", JSON.stringify(servicios));
+	console.log(servicios)
+	res.render('seleccionar-empleado', {
+		userEmpleados: userEmpleados,
+		personaEmpleados: personaEmpleados,
+		selServicios: servicios
+	});
+});
+
+router.get('/admin/cliente', isAuthenticated, async (req, res, next) => {
+	const userCliente = await Usuarios.find({ tipo: "Cliente" });
+	const aux = await Usuarios.find({ tipo: "Cliente" });
+	const ids = [];
+	aux.forEach(auxi => {
+		ids.push(auxi._id);
+	});
+	const personaCliente = await Personas.find({ usuario: { $in: ids } });
+
+	res.render('admin/cliente', {
+		listaUsers: userCliente,
+		listaPersonas: personaCliente
+	});
 });
 
 router.get('/admin/empleado', isAuthenticated, async (req, res, next) => {
@@ -142,9 +233,7 @@ router.get('/admin/empleado', isAuthenticated, async (req, res, next) => {
 	aux.forEach(auxi => {
 		ids.push(auxi._id);
 	});
-	//console.log(ids);
 	const personaEmpleados = await Personas.find({ usuario: { $in: ids } });
-	//console.log(personaEmpleados);
 
 	res.render('admin/empleado', {
 		listaUsers: userEmpleados,
@@ -179,41 +268,81 @@ router.post('/admin/empleado', isAuthenticated, async (req, res, next) => {
 		await unlink(req.file.path);
 	}
 
-	res.render('admin/empleado');
+	const userEmpleados = await Usuarios.find({ tipo: "Empleado" });
+	const aux = await Usuarios.find({ tipo: "Empleado" });
+	const ids = [];
+	aux.forEach(auxi => {
+		ids.push(auxi._id);
+	});
+	const personaEmpleados = await Personas.find({ usuario: { $in: ids } });
+
+	res.render('admin/empleado', {
+		listaUsers: userEmpleados,
+		listaPersonas: personaEmpleados
+	});
 });
+
+router.post('/editar/empleado', isAuthenticated, async (req, res, next) => {
+	const userEmpleados = await Usuarios.find({ tipo: "Empleado" });
+	const aux = await Usuarios.find({ tipo: "Empleado" });
+	const ids = [];
+	aux.forEach(auxi => {
+		ids.push(auxi._id);
+	});
+	const empleados = await Personas.find({ usuario: { $in: ids } });
+	
+	const { nombre, apellido, nacimiento, cedula, telefono, email, password, posicion } = req.body;
+	const actEmpleado = empleados[parseInt(posicion)];
+	const actUsuario = userEmpleados[parseInt(posicion)];
+	actEmpleado.nombre = nombre;
+	actEmpleado.apellido = apellido;
+	actEmpleado.nacimiento = nacimiento;
+	actEmpleado.cedula = cedula;
+	actEmpleado.telefono = telefono;
+	actUsuario.email = email;
+	actUsuario.password = password;
+
+
+	if (req.file) {
+		const result = await Cloudinary.v2.uploader.upload(req.file.path);
+		actEmpleado.urlimage = result.url;
+		actEmpleado.public_id = result.public_id;
+	}
+	actEmpleado.save();
+	actUsuario.save();
+
+	userEmpleados = await Usuarios.find({ tipo: "Empleado" });
+	aux = await Usuarios.find({ tipo: "Empleado" });
+	ids = [];
+	aux.forEach(auxi => {
+		ids.push(auxi._id);
+	});
+	const personaEmpleados = await Personas.find({ usuario: { $in: ids } });
+
+	res.render('admin/empleado', {
+		listaUsers: userEmpleados,
+		listaPersonas: personaEmpleados
+	});
+});
+
+router.post('/eliminar/empleado', isAuthenticated, async (req, res, next) => {
+	const servicios = await Servicios.find({});
+	const { posicion } = req.body;
+	const servicio = servicios[parseInt(posicion)];
+	await servicio.remove();
+	console.log(servicio);
+	const listaservicios = await Servicios.find({});
+	res.render('admin/empleado', {
+		listaServicios: listaservicios
+	});
+});
+
+
 
 router.get('/admin/servicio', isAuthenticated, async (req, res, next) => {
 	const servicios = await Servicios.find({});
 	res.render('admin/servicio', {
 		listaServicios: servicios
-	});
-});
-
-router.post('/admin/servicio/eliminar', isAuthenticated, async (req, res, next) => {
-	const servicios = await Servicios.findOne({ nombre:req.body.id });
-	console.log(servicios);
-	res.render('admin/servicio', {
-		listaServicios: servicios
-	});
-});
-
-router.post('/editar/servicio', isAuthenticated, async (req, res, next) => {
-	const servicios = await Servicios.find({});
-	const { nombre, descripcion, tipo, precio, image, posicion} = req.body;
-	const actServicio = servicios[parseInt(posicion)];
-	actServicio.nombre = nombre;
-	actServicio.descripcion = descripcion;
-	actServicio.tipo = tipo;
-	actServicio.precio = precio;
-	if(req.file){
-		const result = await Cloudinary.v2.uploader.upload(req.file.path);
-		actServicio.urlimage = result.url;
-		actServicio.public_id = result.public_id;	
-	}
-	await actServicio.save();
-	const listaservicios = await Servicios.find({});
-	res.render('admin/servicio', {
-		listaServicios: listaservicios
 	});
 });
 
@@ -224,7 +353,7 @@ router.post('/admin/servicio', isAuthenticated, async (req, res, next) => {
 	const newServicio = new Servicios();
 	newServicio.nombre = nombre;
 	newServicio.descripcion = descripcion;
-	newServicio.tipo = tipo;
+	newServicio.tipo = "Fijo";
 	newServicio.precio = precio;
 	const result = await Cloudinary.v2.uploader.upload(req.file.path);
 	newServicio.urlimage = result.url;
@@ -235,6 +364,37 @@ router.post('/admin/servicio', isAuthenticated, async (req, res, next) => {
 
 	res.render('admin/servicio', {
 		listaServicios: servicios
+	});
+});
+
+router.post('/editar/servicio', isAuthenticated, async (req, res, next) => {
+	const servicios = await Servicios.find({});
+	const { nombre, descripcion, tipo, precio, image, posicion } = req.body;
+	const actServicio = servicios[parseInt(posicion)];
+	actServicio.nombre = nombre;
+	actServicio.descripcion = descripcion;
+	actServicio.precio = precio;
+	if (req.file) {
+		const result = await Cloudinary.v2.uploader.upload(req.file.path);
+		actServicio.urlimage = result.url;
+		actServicio.public_id = result.public_id;
+	}
+	await actServicio.save();
+	const listaservicios = await Servicios.find({});
+	res.render('admin/servicio', {
+		listaServicios: listaservicios
+	});
+});
+
+router.post('/eliminar/servicio', isAuthenticated, async (req, res, next) => {
+	const servicios = await Servicios.find({});
+	const { posicion } = req.body;
+	const servicio = servicios[parseInt(posicion)];
+	await servicio.remove();
+	console.log(servicio);
+	const listaservicios = await Servicios.find({});
+	res.render('admin/servicio', {
+		listaServicios: listaservicios
 	});
 });
